尝试重构TreeView的持久化方式

+++ 自定义控件/uc-Editor.xaml.cs
@@ -293,7 +293,7 @@ namespace 脸滚键盘
                 //btnSaveDoc.Content = "保存■";
                 btnSaveDoc.IsEnabled = true;
             }
-            EditorOperate.ShowValue(words, lbValue);
+            EditorOperate.ShowValue((double)words, lbValue);
         }
